Implement *much* faster `edit_dist()`
The old implementation was accurate and correct in it's output, however
myself and many students noticed an extreme slowdown when comparing
against a large input. For example, a key and observed file of a 50x50
grid was taking upwards of 5 minutes and ~90% CPU utilization on student
machines (and also gradescope).

In research and using my experience from the alignment project in CS
312, I discovered that the use of tuples in the dictionary was causing a
massive slowdown. I believe this to be because of each lookup requiring
a new hash of the tuple, lookup, and allocating a new tuple key for each
access. When asking ChatGPT it said "Since Needleman–Wunsch does O(n*m)
lookups, this turns what should be a fast numeric DP into millions of
Python object allocations and hash lookups."

I had it help me create a new and faster version using a matrix, which
appears to have solved the problem and made the tests that previously
ran incredibly slowly to run in about 5 seconds.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>