Use default of pytest v9
There are no breaking changes in upgrading our pytest base, as far as I
can tell, and all tests appear to be unchanged before and after (aside
from a few bugs that have apparently occured over time as changes have
been made that I fixed).

This fixes a few minor issues, and prevents the deprecation warnings
that are starting to pop up in Python 3.14 with the previous pytest
version.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>